ReadMe
Brief Description.
When i was in 2nd year, i came to know about Machine Learning and wanted to do something with it. After much googling, i found out that, classifying the MNIST datasets is like printing the HELLO WORLD in machine learning arena. So, i thought why not print that now. :)